instrumentation/python/logs/logginginstrumentor/app/requirements.txt
flask
#opentelemetry-exporter-otlp
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-logging
requests
